+++ workflows_definition/RELEASES.md
@@ -59,3 +59,11 @@ Based on:
 - Speakeasy CLI 1.361.1 (2.393.4) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.3.2] workflows_definition
+
+## 2024-08-15 00:25:01
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.368.0 (2.399.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.3.3] workflows_definition